Refactor emoji-related classes and update application configuration for improved clarity and functionality
