+++ src/main/java/oneseoktwojo/ohtalkhae/domain/auth/entity/User.java
@@ -31,4 +31,7 @@ public class User {
     private Role role;
     @CreatedDate
     private LocalDateTime createdAt;
+
+    private String nickname;
+    private String profileImagePath;
 }
