+++ src/main/java/oneseoktwojo/ohtalkhae/domain/friend/repository/FriendRelationRepository.java
@@ -1,7 +1,7 @@
 package oneseoktwojo.ohtalkhae.domain.friend.repository;
 
 import oneseoktwojo.ohtalkhae.domain.friend.entity.FriendRelation;
-import oneseoktwojo.ohtalkhae.domain.profile.User;
+import oneseoktwojo.ohtalkhae.domain.auth.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
