--- src/main/java/oneseoktwojo/ohtalkhae/domain/profile/User.java
@@ -1,22 +1,0 @@
-package oneseoktwojo.ohtalkhae.domain.profile;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String nickname;
-    private String profileImagePath;
-}
-
-
