+++ src/main/java/oneseoktwojo/ohtalkhae/domain/profile/UserProfileService.java
@@ -11,6 +11,8 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
+import oneseoktwojo.ohtalkhae.domain.auth.repository.UserRepository;
+import oneseoktwojo.ohtalkhae.domain.auth.entity.User;
 
 @Service
 @RequiredArgsConstructor
