--- src/main/java/oneseoktwojo/ohtalkhae/domain/profile/UserRepository.java
@@ -1,6 +1,0 @@
-package oneseoktwojo.ohtalkhae.domain.profile;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
