+++ src/main/resources/application.yml
@@ -26,6 +26,6 @@ spring:
     properties:
       hibernate:
         format_sql: true
-  file:
-    upload:
-      path: ./upload/profile/
+file:
+  upload:
+    path: ./upload/profile/
